refactor: replace FileValidatorManager and InMemoryValidatorManager with DefaultValidatorManager
Consolidate two nearly-identical ValidatorManager implementations into a single
DefaultValidatorManager that accepts an Optional<KeystoreFileManager>. When present the manager
persists keystores to disk (file-backed mode); when empty it operates in-memory only, matching the
previous InMemoryValidatorManager behaviour. KeyManagerApiRoute now always constructs a
DefaultValidatorManager, passing Optional.empty() when skipKeystoreStorage is enabled.

Tests from InMemoryValidatorManagerTest are merged into DefaultValidatorManagerTest, covering both
modes and exception paths. The deleted files are no longer referenced anywhere.
