+++ signing/src/main/java/tech/pegasys/web3signer/signing/DefaultValidatorManager.java
@@ -13,18 +13,36 @@
 package tech.pegasys.web3signer.signing;
 
 import java.io.IOException;
-import java.io.UncheckedIOException;
+import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 
 import org.apache.tuweni.bytes.Bytes;
 
-public class FileValidatorManager implements ValidatorManager {
+/// Default [`ValidatorManager`] implementation that supports both file-backed and in-memory-only
+/// validator management.
+///
+/// When a [`KeystoreFileManager`] is supplied, validators are persisted to disk before being
+/// activated in memory, so they survive a restart. When no file manager is supplied (i.e.
+/// `skipKeystoreStorage` is enabled), validators exist only in memory and are lost when
+/// Web3Signer stops.
+///
+/// In both modes the in-memory state is always the source of truth for active signing: additions
+/// register the signer in memory _after_ a successful file write (if applicable), and deletions
+/// remove the signer from memory _before_ deleting any files.
+public class DefaultValidatorManager implements ValidatorManager {
 
   private final ArtifactSignerProvider signerProvider;
-  private final KeystoreFileManager keystoreFileManager;
+  private final Optional<KeystoreFileManager> keystoreFileManager;
 
-  public FileValidatorManager(
-      final ArtifactSignerProvider signerProvider, final KeystoreFileManager keystoreFileManager) {
+  /// Creates a new `DefaultValidatorManager`.
+  ///
+  /// @param signerProvider the registry used to activate and deactivate signers in memory
+  /// @param keystoreFileManager when present, keystore files are written on import and deleted on
+  ///     removal; when empty, the manager operates in in-memory-only mode and imported validators
+  ///     are **not** persisted across restarts
+  public DefaultValidatorManager(
+      final ArtifactSignerProvider signerProvider,
+      final Optional<KeystoreFileManager> keystoreFileManager) {
     this.signerProvider = signerProvider;
     this.keystoreFileManager = keystoreFileManager;
   }
@@ -35,7 +53,9 @@ public class FileValidatorManager implements ValidatorManager {
       // Remove active key from memory first, will stop any further signing with this key
       signerProvider.removeSigner(publicKey.toHexString()).get();
       // Then, delete the corresponding keystore files
-      keystoreFileManager.deleteKeystoreFiles(publicKey.toHexString());
+      if (keystoreFileManager.isPresent()) {
+        keystoreFileManager.get().deleteKeystoreFiles(publicKey.toHexString());
+      }
     } catch (InterruptedException e) {
       Thread.currentThread().interrupt();
       throw new IllegalStateException("Unable to delete validator", e);
@@ -49,16 +69,16 @@ public class FileValidatorManager implements ValidatorManager {
       final BlsArtifactSigner signer, final KeystoreFileRecord keystoreFileRecord) {
     try {
       // write keystore to file - allows to bail out in case of failure
-      keystoreFileManager.createKeystoreFiles(keystoreFileRecord);
+      if (keystoreFileManager.isPresent()) {
+        keystoreFileManager.get().createKeystoreFiles(keystoreFileRecord);
+      }
       // Then, add it in memory to make it available for signing ...
       signerProvider.addSigner(signer).get();
     } catch (InterruptedException e) {
       Thread.currentThread().interrupt();
       throw new IllegalStateException("Unable to add validator", e);
-    } catch (ExecutionException e) {
+    } catch (ExecutionException | IOException e) {
       throw new IllegalStateException("Unable to add validator", e);
-    } catch (IOException e) {
-      throw new UncheckedIOException(e);
     }
   }
 }
